requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
crispy-bootstrap5==0.6
Django==4.0.2
django-crispy-forms==1.14.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
Pillow==9.0.1
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
